Suggested changes from review
‘as it can be styled in a consistent way and also can be used to change
the display by assistive technologies.’ - not quite sure what you are
aiming at with this one?
Lists: ‘form elements’ - might be a bug bear of mine, should we be
hinting that placing form elements in a list is ok?
Definition lists: I think this might be an item for discussion as you
only describe one use.
Images: ‘help guiding users’ - any examples how?

Overall I am not sure of the value of this section. It is good at
explaining how to mark up basic HTML, but how much accessibility
information does it impart?