Suggested changes from review
Visual clarity: ‘pages with many components’ is there an
internationalisation issue here… asian web design tends to go ott on
visual clutter and that generally (usability) tests well?
Font size: ‘It is not possible to zoom text set in pixels independently
from the rest of the page’ - is this true? Or more particularly… what
does it mean?!